requirements.txt
Django==3.2.*
djangorestframework==3.12.*
gunicorn==20.0.*
Inject==4.3.*
pydantic==1.8.*
mysqlclient==2.1.*
redis==4.3.*
celery==5.2.*
beautifulsoup4==4.11.*
requests==2.28.*
